refactor: split LogLayer.basic.test.ts into smaller files
Extract log level management (191 lines) and tagged template syntax
(247 lines) tests into their own files. Reduces basic test file from
2021 lines (61KB) to 1622 lines (48KB) to avoid biome linter OOM.
