+++ packages/core/loglayer/src/__tests__/LogLayer.basic.test.ts
@@ -1618,404 +1618,4 @@ describe("LogLayer basic functionality", () => {
       expect(obj.fromRoot).toBe("yes");
     });
   });
-
-  describe("log level management", () => {
-    it("should enable individual log levels", () => {
-      const log = getLogger();
-      const genericLogger = log.getLoggerInstance("console") as TestLoggingLibrary;
-
-      // Disable all log levels first
-      log.disableLogging();
-
-      // Verify logging is disabled
-      log.info("test info");
-      expect(genericLogger.popLine()).not.toBeDefined();
-
-      // Enable just the info level
-      log.enableIndividualLevel(LogLevel.info);
-
-      // Info should now work
-      log.info("test info");
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          level: LogLevel.info,
-          data: ["test info"],
-        }),
-      );
-
-      // But other levels should still be disabled
-      log.warn("test warn");
-      expect(genericLogger.popLine()).not.toBeDefined();
-
-      log.error("test error");
-      expect(genericLogger.popLine()).not.toBeDefined();
-    });
-
-    it("should disable individual log levels", () => {
-      const log = getLogger();
-      const genericLogger = log.getLoggerInstance("console") as TestLoggingLibrary;
-
-      // Start with all levels enabled
-      log.enableLogging();
-
-      // Disable just the warn level
-      log.disableIndividualLevel(LogLevel.warn);
-
-      // Warn should be disabled
-      log.warn("test warn");
-      expect(genericLogger.popLine()).not.toBeDefined();
-
-      // Other levels should still work
-      log.info("test info");
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          level: LogLevel.info,
-          data: ["test info"],
-        }),
-      );
-
-      log.error("test error");
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          level: LogLevel.error,
-          data: ["test error"],
-        }),
-      );
-    });
-
-    it("should set minimum log level", () => {
-      const log = getLogger();
-      const genericLogger = log.getLoggerInstance("console") as TestLoggingLibrary;
-
-      // Set minimum level to warn
-      log.setLevel(LogLevel.warn);
-
-      // Info and debug should be disabled
-      log.info("test info");
-      expect(genericLogger.popLine()).not.toBeDefined();
-
-      log.debug("test debug");
-      expect(genericLogger.popLine()).not.toBeDefined();
-
-      // Warn, error and fatal should be enabled
-      log.warn("test warn");
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          level: LogLevel.warn,
-          data: ["test warn"],
-        }),
-      );
-
-      log.error("test error");
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          level: LogLevel.error,
-          data: ["test error"],
-        }),
-      );
-
-      log.fatal("test fatal");
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          // console logger does not have a fatal level
-          level: LogLevel.error,
-          data: ["test fatal"],
-        }),
-      );
-    });
-
-    it("should correctly report if a level is enabled", () => {
-      const log = getLogger();
-
-      // Start with all levels enabled
-      log.enableLogging();
-
-      expect(log.isLevelEnabled(LogLevel.info)).toBe(true);
-      expect(log.isLevelEnabled(LogLevel.warn)).toBe(true);
-      expect(log.isLevelEnabled(LogLevel.error)).toBe(true);
-      expect(log.isLevelEnabled(LogLevel.debug)).toBe(true);
-      expect(log.isLevelEnabled(LogLevel.trace)).toBe(true);
-      expect(log.isLevelEnabled(LogLevel.fatal)).toBe(true);
-
-      // Disable specific levels
-      log.disableIndividualLevel(LogLevel.debug);
-      log.disableIndividualLevel(LogLevel.info);
-
-      expect(log.isLevelEnabled(LogLevel.info)).toBe(false);
-      expect(log.isLevelEnabled(LogLevel.warn)).toBe(true);
-      expect(log.isLevelEnabled(LogLevel.error)).toBe(true);
-      expect(log.isLevelEnabled(LogLevel.debug)).toBe(false);
-      expect(log.isLevelEnabled(LogLevel.trace)).toBe(true);
-      expect(log.isLevelEnabled(LogLevel.fatal)).toBe(true);
-
-      // Set minimum level to error
-      log.setLevel(LogLevel.error);
-
-      expect(log.isLevelEnabled(LogLevel.info)).toBe(false);
-      expect(log.isLevelEnabled(LogLevel.warn)).toBe(false);
-      expect(log.isLevelEnabled(LogLevel.error)).toBe(true);
-      expect(log.isLevelEnabled(LogLevel.debug)).toBe(false);
-      expect(log.isLevelEnabled(LogLevel.trace)).toBe(false);
-      expect(log.isLevelEnabled(LogLevel.fatal)).toBe(true);
-    });
-
-    it("should chain log level methods", () => {
-      const log = getLogger();
-      const genericLogger = log.getLoggerInstance("console") as TestLoggingLibrary;
-
-      // Chain multiple operations
-      log.disableLogging().enableIndividualLevel(LogLevel.error).enableIndividualLevel(LogLevel.fatal);
-
-      // Debug should be disabled
-      log.debug("test debug");
-      expect(genericLogger.popLine()).not.toBeDefined();
-
-      // Error should be enabled
-      log.error("test error");
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          level: LogLevel.error,
-          data: ["test error"],
-        }),
-      );
-
-      // Chain with other methods
-      log.setLevel(LogLevel.warn).withMetadata({ test: "data" }).warn("test with chain");
-
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          level: LogLevel.warn,
-          data: [{ test: "data" }, "test with chain"],
-        }),
-      );
-    });
-  });
-
-  describe("tagged template syntax", () => {
-    it("should support tagged template syntax on info", () => {
-      const log = getLogger();
-      const genericLogger = log.getLoggerInstance("console") as TestLoggingLibrary;
-      const userId = "123";
-
-      log.info`User ${userId} logged in`;
-
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          level: LogLevel.info,
-          data: ["User 123 logged in"],
-        }),
-      );
-    });
-
-    it("should support tagged template syntax on all log levels", () => {
-      const log = getLogger();
-      const genericLogger = log.getLoggerInstance("console") as TestLoggingLibrary;
-      const userId = "123";
-
-      log.info`User ${userId} logged in`;
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({ level: LogLevel.info, data: ["User 123 logged in"] }),
-      );
-
-      log.warn`User ${userId} logged out`;
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({ level: LogLevel.warn, data: ["User 123 logged out"] }),
-      );
-
-      log.error`User ${userId} failed login`;
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({ level: LogLevel.error, data: ["User 123 failed login"] }),
-      );
-
-      log.debug`User ${userId} session checked`;
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({ level: LogLevel.debug, data: ["User 123 session checked"] }),
-      );
-
-      log.trace`User ${userId} entering`;
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({ level: LogLevel.trace, data: ["User 123 entering"] }),
-      );
-
-      log.fatal`User ${userId} system crash`;
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({ level: LogLevel.error, data: ["User 123 system crash"] }),
-      );
-    });
-
-    it("should support multiple interpolations in tagged template", () => {
-      const log = getLogger();
-      const genericLogger = log.getLoggerInstance("console") as TestLoggingLibrary;
-
-      log.info`User ${"alice"} performed ${"login"} at ${new Date(0).toISOString()}`;
-
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          level: LogLevel.info,
-          data: ["User alice performed login at 1970-01-01T00:00:00.000Z"],
-        }),
-      );
-    });
-
-    it("should support tagged template with context", () => {
-      const log = getLogger();
-      const genericLogger = log.getLoggerInstance("console") as TestLoggingLibrary;
-      const requestId = "req-123";
-
-      log.withContext({ requestId }).info`Processing request ${requestId}`;
-
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          level: LogLevel.info,
-          data: [{ requestId: "req-123" }, "Processing request req-123"],
-        }),
-      );
-    });
-
-    it("should support tagged template with metadata", () => {
-      const log = getLogger();
-      const genericLogger = log.getLoggerInstance("console") as TestLoggingLibrary;
-
-      log.withMetadata({ duration: 150 }).warn`Request took ${150}ms`;
-
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          level: LogLevel.warn,
-          data: [{ duration: 150 }, "Request took 150ms"],
-        }),
-      );
-    });
-
-    it("should support tagged template with error", () => {
-      const log = getLogger();
-      const genericLogger = log.getLoggerInstance("console") as TestLoggingLibrary;
-      const err = new Error("timeout");
-
-      log.withError(err).error`Request failed: ${err.message}`;
-
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          level: LogLevel.error,
-          data: [{ err }, "Request failed: timeout"],
-        }),
-      );
-    });
-
-    it("should handle null values in tagged template", () => {
-      const log = getLogger();
-      const genericLogger = log.getLoggerInstance("console") as TestLoggingLibrary;
-      const nullValue = null;
-
-      log.info`User ${nullValue} logged in`;
-
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          level: LogLevel.info,
-          data: ["User null logged in"],
-        }),
-      );
-    });
-
-    it("should handle undefined values in tagged template", () => {
-      const log = getLogger();
-      const genericLogger = log.getLoggerInstance("console") as TestLoggingLibrary;
-      const undefinedValue = undefined;
-
-      log.info`User ${undefinedValue} logged in`;
-
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          level: LogLevel.info,
-          data: ["User undefined logged in"],
-        }),
-      );
-    });
-
-    it("should handle object values in tagged template", () => {
-      const log = getLogger();
-      const genericLogger = log.getLoggerInstance("console") as TestLoggingLibrary;
-      const user = { name: "Alice" };
-
-      log.info`User ${user} logged in`;
-
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          level: LogLevel.info,
-          data: ["User [object Object] logged in"],
-        }),
-      );
-    });
-
-    it("should preserve template spaces in tagged template", () => {
-      const log = getLogger();
-      const genericLogger = log.getLoggerInstance("console") as TestLoggingLibrary;
-
-      log.info`Hello   world   with   spaces`;
-
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          level: LogLevel.info,
-          data: ["Hello   world   with   spaces"],
-        }),
-      );
-    });
-
-    it("should support tagged template with prefix", () => {
-      const log = getLogger().withPrefix("[APP]");
-      const genericLogger = log.getLoggerInstance("console") as TestLoggingLibrary;
-
-      log.info`User ${"123"} logged in`;
-
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          level: LogLevel.info,
-          data: ["[APP] User 123 logged in"],
-        }),
-      );
-    });
-
-    it("should support tagged template with groups", () => {
-      const log = getLogger();
-      const genericLogger = log.getLoggerInstance("console") as TestLoggingLibrary;
-
-      log.withMetadata({ test: "data" }).withGroup("auth").info`User ${"123"} logged in`;
-
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          level: LogLevel.info,
-          data: [{ test: "data" }, "User 123 logged in"],
-        }),
-      );
-    });
-
-    it("should still work with regular call syntax alongside tagged template", () => {
-      const log = getLogger();
-      const genericLogger = log.getLoggerInstance("console") as TestLoggingLibrary;
-
-      log.info("Regular call");
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          level: LogLevel.info,
-          data: ["Regular call"],
-        }),
-      );
-
-      const userId = "456";
-      log.info`Tagged template ${userId}`;
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          level: LogLevel.info,
-          data: ["Tagged template 456"],
-        }),
-      );
-
-      log.info("Another regular call", "with multiple args");
-      expect(genericLogger.popLine()).toStrictEqual(
-        expect.objectContaining({
-          level: LogLevel.info,
-          data: ["Another regular call", "with multiple args"],
-        }),
-      );
-    });
-  });
 });
